pyproject.toml
[tool.poetry]
name = "voice-ai-agent"
version = "0.1.0"
description = "Voice AI Agent that handles Twilio calls"
authors = ["Tim Siwula <siwulactim@gmail.com>"]
readme = "README.md"
packages = [
  { include = "src", from = "." }
]

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
fastapi = "^0.95.0"
uvicorn = { version = "^0.23.0", extras = ["standard"] }
twilio = "^8.0.0"
python-dotenv = "^1.0.0"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
httpx = "^0.26.0"
pytest-dotenv = "^0.5.2"

[tool.poetry.scripts]
start = "src.server:main"

[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"
